# packages/artifact-host/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Standalone Drop-compatible artifact host",
    "version": "0.1.0",
    "description": "Independent reference behavior targeting Drop e924002df3e0cd88c6a18c03e45ecc4c2c88c14a after URL and bearer-auth transport configuration. No WorkOS implementation or branding is included. Phase 2 comment/review routes are not implemented. Uploads start private. Legacy clients may omit guards and use serialized last-accepted writes. Pi uses quoted-integer If-Match and durable Idempotency-Key for upload/replacement. Matching retries replay the original result even after newer changes. Keys are retained as bounded tombstones, including after deletion. Reuse with different content is 409. A stale If-Match is 412. Visibility and rename do not advance the content version. Limits: 25 MiB file bytes, 26 MiB multipart envelope, 256 files, 1024 UTF-8 bytes/path, 16 path segments, 1000 title characters, 1000 artifacts, 100 versions/artifact, 1 GiB aggregate assets/history, 10000 retained operations, 4 concurrent upload bodies, 600 owner requests/minute, 3000 viewing auth records. Limits reject before mutation without eviction. All executable file responses carry a CSP sandbox, including direct HTML and SVG navigation."
  },
  "security": [{ "ownerBearer": [] }],
  "paths": {
...
    "/api/links/{slug}/file": {
      "parameters": [{ "$ref": "#/components/parameters/Slug" }],
      "post": {
        "operationId": "replace",
        "parameters": [
          { "$ref": "#/components/parameters/Idempotency" },
          { "$ref": "#/components/parameters/IfMatch" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/Multipart" },
        "responses": {
          "200": { "$ref": "#/components/responses/Published" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ownerBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Single configured owner. Viewer cookies, capabilities and email headers never establish this authority."
      },
      "viewerCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "artifact_viewer",
        "description": "Read-only private viewer session. Never accepted by owner APIs."
      }
    },
    "parameters": {
      "Slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "pattern": "^[a-z0-9]{8,16}$" }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": false,
        "schema": { "type": "string", "pattern": "^\"[0-9]+\"$" },
        "description": "Quoted integer content version. Upload accepts 0. Pi always uses guards."
      },
      "Idempotency": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": { "type": "string", "pattern": "^[A-Za-z0-9_-]{1,128}$" }
      }
    },
    "requestBodies": {
      "Multipart": {
        "required": true,
        "content": {
          "multipart/form-data": {
            "schema": {
              "type": "object",
              "required": ["files", "paths"],
              "properties": {
                "files": { "type": "array", "maxItems": 256, "items": { "type": "string", "format": "binary" } },
                "paths": {
                  "type": "string",
                  "description": "JSON array of safe relative paths in the same order as repeated files. A lone root HTML is normalized to index.html, companion asset paths are preserved."
                },
                "title": { "type": "string", "maxLength": 1000 }
              }
            }
          }
        }
      }
    },
...
    "responses": {
      "Published": {
        "description": "Upload or stable-slug replacement/restore. ok is present on replace/restore. Version and visibility are additive.",
        "headers": { "ETag": { "schema": { "type": "string" }, "description": "Quoted integer current_version" } },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["slug", "url", "current_version", "is_public"],
              "properties": {
                "ok": { "const": true },
                "slug": { "type": "string" },
                "url": { "type": "string", "format": "uri" },
                "current_version": { "type": "integer" },
                "is_public": { "type": "boolean" }
              }
            }
          }
        }
      },
...
      "Error": {
        "description": "400 invalid input, 401 missing/invalid authority, 403 origin denied, 404 absent, 409 key mismatch, 412 version conflict, 413 body/file limit, 415 content type, 429 rate/concurrency/auth limit, 500 transaction failure, 507 storage/history/operation limit. No secrets in errors.",
        "content": {
          "application/json": {
            "schema": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } } }
          }
        }
      }
    }
  }
}